# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "Laravel API OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Laravel API Server"
        }
    ],
    "paths": {
...
        "/sensors": {
            "get": {
                "tags": [
                    "sensors"
                ],
                "summary": "Get All sensor",
                "operationId": "cce7ed93f1ada7ef3cd47c0aceac057f",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "bearerAuth"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "allowEmptyValue": true,
                        "example": "L222"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "allowEmptyValue": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sensor retrieve succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResult"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Sensor"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not Allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "sensors"
                ],
                "summary": "Add new sensor",
                "operationId": "632e0e6a4ef0dfbbf076a6e80ec67cce",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "bearerAuth"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Sensor data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_id",
                                    "device_addr"
                                ],
                                "properties": {
                                    "device_addr": {
                                        "type": "string",
                                        "example": "fe:80:70:40:70"
                                    },
                                    "room_id": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sensor created succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sensor"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Not Allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sensors/{sensorId}": {
            "get": {
                "tags": [
                    "sensors"
                ],
                "summary": "Update sensor",
                "operationId": "6af62a07f3aa19eb8b7d5b16cb83176e",
                "parameters": [
                    {
                        "name": "sensorId",
                        "in": "path",
                        "description": "Id Of selected sensor",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sensor created succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sensor"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not Allowed"
                    }
                }
            },
            "put": {
                "tags": [
                    "sensors"
                ],
                "summary": "Update sensor",
                "operationId": "06a37ae0b7dbfa05b14274600068af92",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "bearerAuth"
                        }
                    },
                    {
                        "name": "sensorId",
                        "in": "path",
                        "description": "Id Of selected sensor",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Sensor data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_id",
                                    "device_addr"
                                ],
                                "properties": {
                                    "device_addr": {
                                        "type": "string",
                                        "example": "fe:80:70:40:70"
                                    },
                                    "room_id": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sensor created succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sensor"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Not Allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "sensors"
                ],
                "summary": "Update sensor",
                "operationId": "5d8347af8262065673dc4fc9023d182c",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "bearerAuth"
                        }
                    },
                    {
                        "name": "sensorId",
                        "in": "path",
                        "description": "Id Of selected sensor",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sensor deleted succesfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Not Allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sensors/{sensorId}/qrcode": {
            "get": {
                "tags": [
                    "sensors"
                ],
                "summary": "Get Code code",
                "operationId": "9897f168ce2e6ce4a6ebe6e5dfaa9ed5",
                "parameters": [
                    {
                        "name": "sensorId",
                        "in": "path",
                        "description": "Id Of selected sensor",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get sensor qrCode succesfully (return png images)"
                    },
                    "403": {
                        "description": "Not Allowed"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginatedResult": {
                "title": "Pagination exemple",
                "properties": {
                    "total": {
                        "type": "integer"
                    },
                    "current_page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "next_page_url": {
                        "type": "string"
                    },
                    "prev_page_url": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
...
            "Sensor": {
                "title": "schema of sensor",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "last_message": {
                        "type": "string"
                    },
                    "created_by": {
                        "type": "string"
                    },
                    "room_id": {
                        "type": "integer"
                    },
                    "id_hex": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}